Fixes the registry (#99)
Our registry pattern is far from idea and thus we haven't used it consistent. This PR aims to resolve this by copying the registry pattern used in Pythia (as it is much nicer) and then registering all current measures and sensors and using the register to query for those instead of getattr as this is a more clearly extendable pattern (as compared to monkey patching).